+++ api/tests/certification/session/integration/infrastructure/utils/pdf/attendance-sheet-EN_expected.pdf
