+++ api/tests/certification/session/integration/infrastructure/utils/pdf/attendance-sheet-with-1-page_expected.pdf
