+++ api/tests/certification/session/integration/infrastructure/utils/pdf/attendance-sheet-with-external-id_expected.pdf
