+++ api/tests/certification/session/integration/infrastructure/utils/pdf/attendance-sheet-with-optional-value_expected.pdf
