+++ api/tests/certification/session/integration/infrastructure/utils/pdf/attendance-sheet-with-truncated-information_expected.pdf
