+++ api/tests/certification/session/integration/infrastructure/utils/pdf/attendance-sheet_expected.pdf
